+++ gamevault/Converter/InverseNullConverter.cs
@@ -12,7 +12,7 @@ namespace gamevault.Converter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return (value == null);
+           return (value == null);
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)