+++ gamevault/Models/AppInfo.cs
@@ -17,6 +17,8 @@ namespace gamevault.Models
         ServerUrl,
         LibStartup,
         AutoExtract,
+        AutoInstallPortable,
+        AutoDeletePortable,
         ExtractionFinished,
         RunningInTrayMessage,
         DownloadLimit,
@@ -28,7 +30,9 @@ namespace gamevault.Models
         Phalcode2,
         Theme,
         ExtractionPassword,
-        DownloadProgress
+        DownloadProgress,
+        InstalledGamesRows,
+        CreateDesktopShortcut
     }
     public static class AppFilePath
     {
@@ -58,6 +62,7 @@ namespace gamevault.Models
     }
     public static class Globals
     {
-        internal static string[] SupportedExecutables = new string[] { "EXE", "BAT", "COM", "CMD", "INF", "IPA", "OSX", "PIF", "RUN", "WSH", "LNK", "SH" };
+        internal static string[] SupportedExecutables = new string[] { "EXE", "BAT", "COM", "CMD", "INF", "IPA", "OSX", "PIF", "RUN", "WSH", "LNK", "SH", "MSI" };
+        internal static string[] SupportedArchives = new string[] { ".7z", ".xz", ".bz2", ".gz", ".tar", ".zip", ".wim", ".ar", ".arj", ".cab", ".chm", ".cpio", ".cramfs", ".dmg", ".ext", ".fat", ".gpt", ".hfs", ".ihex", ".iso", ".lzh", ".lzma", ".mbr", ".msi", ".nsis", ".ntfs", ".qcow2", ".rar", ".rpm", ".squashfs", ".udf", ".uefi", ".vdi", ".vhd", ".vmdk", ".wim", ".xar", ".z" };
     }
 }