+++ gamevault/UserControls/GameDownloadUserControl.xaml.cs
@@ -4,14 +4,11 @@ using gamevault.Models;
 using gamevault.ViewModels;
 using MahApps.Metro.Controls;
 using MahApps.Metro.Controls.Dialogs;
-using Microsoft.Toolkit.Uwp.Notifications;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Net.Http.Headers;
-using System.Runtime;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
@@ -198,7 +195,8 @@ namespace gamevault.UserControls
 
                 if (downloadRetryTimer.Data != "error")
                 {
-                    ToastMessageHelper.CreateToastMessage("Notification", $"Download of {ViewModel.Game.Title} Failed");
+                    if (!App.Instance.IsWindowActiveAndControlInFocus(MainControl.Downloads))
+                        ToastMessageHelper.CreateToastMessage("Download Failed", ViewModel.Game.Title, $"{AppFilePath.ImageCache}/gbox/{ViewModel.Game.ID}.{ViewModel.Game.BoxImage.ID}");
                 }
                 downloadRetryTimer.Start();
 
@@ -308,7 +306,10 @@ namespace gamevault.UserControls
                 Directory.CreateDirectory(ViewModel.InstallPath);
             }
             MainWindowViewModel.Instance.Library.GetGameInstalls().AddSystemFileWatcher(ViewModel.InstallPath);
-            ToastMessageHelper.CreateToastMessage("Notification", $"Download of {ViewModel.Game.Title} Complete");
+
+            if (!App.Instance.IsWindowActiveAndControlInFocus(MainControl.Downloads))
+                ToastMessageHelper.CreateToastMessage("Download Complete", ViewModel.Game.Title, $"{AppFilePath.ImageCache}/gbox/{ViewModel.Game.ID}.{ViewModel.Game.BoxImage.ID}");
+
             if (SettingsViewModel.Instance.AutoExtract)
             {
                 App.Current.Dispatcher.Invoke((Action)async delegate
@@ -345,18 +346,12 @@ namespace gamevault.UserControls
 
         private string CalculateTimeLeft(long? totalFileSize, long totalBytesRead, double tspanMilliseconds)
         {
-
-            //double tspanSeconds = tspanMilliseconds / 1000; // Convert milliseconds to seconds
-            //Debug.WriteLine($"{totalBytesRead / 100000}/{totalFileSize / 100000} - {tspanSeconds}");
-            //var averageSpeed = totalBytesRead / tspanSeconds;
             double timeLeftSeconds = (double)((totalFileSize - totalBytesRead) / downloadSpeedCalc.GetCurrentSpeed());
-
             TimeSpan t = TimeSpan.FromSeconds(0);
             if (timeLeftSeconds > 0 && !double.IsInfinity(timeLeftSeconds) && !double.IsNaN(timeLeftSeconds))
             {
                 t = TimeSpan.FromSeconds(timeLeftSeconds);
             }
-            //Debug.WriteLine(string.Format("{0:00}:{1:00}:{2:00}", ((int)t.TotalHours), t.Minutes, t.Seconds) + "\n");
             return string.Format("{0:00}:{1:00}:{2:00}", ((int)t.TotalHours), t.Minutes, t.Seconds);
         }
 
@@ -368,6 +363,12 @@ namespace gamevault.UserControls
 
         public async Task DeleteFile(bool confirm = true)
         {
+            if (IsDownloadActive)
+            {
+                MainWindowViewModel.Instance.AppBarText = "Can not delete during the download, extraction, installing process";
+                return;
+            }
+
             bool doDelete = true;
 
             if (confirm)
@@ -388,6 +389,7 @@ namespace gamevault.UserControls
 
                     DownloadsViewModel.Instance.DownloadedGames.Remove(this);
 
+                    //Delete Installation Directory if it is empty (Because GV should not create a filewatcher for it anymore)
                     if (Directory.Exists(ViewModel.InstallPath) && !Directory.EnumerateFileSystemEntries(ViewModel.InstallPath).Any())
                     {
                         Directory.Delete(ViewModel.InstallPath);
@@ -480,10 +482,23 @@ namespace gamevault.UserControls
                 Preferences.Set(AppConfigKey.ExtractionFinished, "1", $"{m_DownloadPath}\\Extract\\gamevault-metadata");
                 ViewModel.State = "Extracted";
                 uiBtnExtract.Text = "Re-Extract";
-                uiBtnInstall.IsEnabled = true;
+
                 ViewModel.InstallationStepperProgress = 1;
                 ViewModel.ExtractionUIVisibility = System.Windows.Visibility.Hidden;
-                ToastMessageHelper.CreateToastMessage("Notification", $"Extraction of {ViewModel.Game.Title} Complete");
+
+                if (!App.Instance.IsWindowActiveAndControlInFocus(MainControl.Downloads))
+                    ToastMessageHelper.CreateToastMessage("Extraction Complete", ViewModel.Game.Title, $"{AppFilePath.ImageCache}/gbox/{ViewModel.Game.ID}.{ViewModel.Game.BoxImage.ID}");
+
+                if (SettingsViewModel.Instance.AutoInstallPortable && (ViewModel.Game.Type == GameType.WINDOWS_PORTABLE || ViewModel.Game.Type == GameType.LINUX_PORTABLE))
+                {
+                    await Task.Delay(1000);//Just to be sure the extraction stream is closed and the files are ready to copy
+                    await Install();
+                }
+                else
+                {
+                    uiBtnInstall.IsEnabled = true;
+                }
+
                 UpdateDataSizeUI();
             }
             else
@@ -508,7 +523,8 @@ namespace gamevault.UserControls
                 else
                 {
                     ViewModel.State = "Something went wrong during extraction";
-                    ToastMessageHelper.CreateToastMessage("Notification", $"Extraction of {ViewModel.Game.Title} Failed");
+                    if (!App.Instance.IsWindowActiveAndControlInFocus(MainControl.Downloads))
+                        ToastMessageHelper.CreateToastMessage("Extraction Failed", ViewModel.Game.Title, $"{AppFilePath.ImageCache}/gbox/{ViewModel.Game.ID}.{ViewModel.Game.BoxImage.ID}");
                 }
                 ViewModel.ExtractionUIVisibility = System.Windows.Visibility.Hidden;
             }
@@ -553,9 +569,14 @@ namespace gamevault.UserControls
                 uiCbSetupExecutable.ItemsSource = null;
             }
         }
-        private async void Install_Click(object sender, RoutedEventArgs e)
+        private async void Install_Click(object s, RoutedEventArgs e)
         {
-            ((FrameworkElement)sender).IsEnabled = false;
+            await Install();
+        }
+        private async Task Install()
+        {
+            uiBtnInstallPortable.IsEnabled = false;
+            uiBtnInstallSetup.IsEnabled = false;
             uiBtnExtract.IsEnabled = false;
             try
             {
@@ -577,7 +598,6 @@ namespace gamevault.UserControls
                         if (!Directory.Exists(ViewModel.InstallPath))
                         {
                             Directory.CreateDirectory(ViewModel.InstallPath);
-                            //MainWindowViewModel.Instance.Installs.AddSystemFileWatcher(ViewModel.InstallPath);
                         }
                         else if (Directory.Exists($"{ViewModel.InstallPath}\\Files"))
                         {
@@ -589,7 +609,10 @@ namespace gamevault.UserControls
                 });
                 uiBtnInstall.IsEnabled = false;
                 uiProgressRingInstall.IsActive = false;
-                ((FrameworkElement)sender).IsEnabled = true;
+
+                uiBtnInstallPortable.IsEnabled = true;
+                uiBtnInstallSetup.IsEnabled = true;
+
                 ViewModel.State = "Downloaded";
                 uiBtnExtract.Text = "Extract";
                 if (error)
@@ -601,6 +624,12 @@ namespace gamevault.UserControls
                     MainWindowViewModel.Instance.AppBarText = $"Successfully installed '{ViewModel.Game.Title}'";
                     ViewModel.InstallationStepperProgress = 2;
                     ViewModel.State = "Installed";
+
+                    //Auto delete files of portable games after successful installation
+                    if (SettingsViewModel.Instance.AutoDeletePortableGameFiles)
+                    {
+                        await DeleteFile(false);
+                    }
                 }
             }
             else if (ViewModel.Game.Type == GameType.WINDOWS_SETUP)
@@ -643,11 +672,26 @@ namespace gamevault.UserControls
                 {
                     MainWindowViewModel.Instance.AppBarText = $"Could not find executable '{setupEexecutable}'";
                 }
-                 ((FrameworkElement)sender).IsEnabled = true;
+                uiBtnInstallPortable.IsEnabled = true;
+                uiBtnInstallSetup.IsEnabled = true;
             }
             uiInstallOptions.Visibility = System.Windows.Visibility.Collapsed;
             uiProgressRingInstall.IsActive = false;
             uiBtnExtract.IsEnabled = true;
+            if (ViewModel.CreateShortcut == true)
+            {
+                await Task.Delay(1000);
+                var game = InstallViewModel.Instance.InstalledGames.Where(g => g.Key.ID == ViewModel.Game.ID).FirstOrDefault();
+                if (game.Key == null)
+                    return;
+
+                if (!GameSettingsUserControl.TryPrepareLaunchExecutable(game.Value))
+                {
+                    MainWindowViewModel.Instance.AppBarText = $"Can not create shortcut. No valid Executable found";
+                    return;
+                }
+                await DesktopHelper.CreateShortcut(game.Key, Preferences.Get(AppConfigKey.Executable, $"{game.Value}\\gamevault-exec"), false);
+            }
         }
 
         private void CopyInstallPathToClipboard_Click(object sender, System.Windows.Input.MouseButtonEventArgs e)
@@ -658,6 +702,29 @@ namespace gamevault.UserControls
                 MainWindowViewModel.Instance.AppBarText = "Copied Installation Directory to Clipboard";
             }
             catch { }
+        }
+
+        private void ContinueOverwriteGameType_Click(object sender, RoutedEventArgs e)
+        {
+            if (uiCbOverwriteGameType.SelectedValue != null)
+            {
+                var temp = ViewModel.Game;
+                temp.Type = (GameType)uiCbOverwriteGameType.SelectedValue;
+                ViewModel.Game = null;
+                ViewModel.Game = temp;
+            }
+            else
+            {
+                MainWindowViewModel.Instance.AppBarText = "No gametype selected for overwriting";
+            }
+        }
+
+        private void InitOverwriteGameType_Click(object sender, System.Windows.Input.MouseButtonEventArgs e)
+        {
+            var temp = ViewModel.Game;
+            temp.Type = GameType.UNDETECTABLE;
+            ViewModel.Game = null;
+            ViewModel.Game = temp;
         }
     }
 }