+++ gamevault/ViewModels/AdminConsoleViewModel.cs
@@ -10,12 +10,18 @@ namespace gamevault.ViewModels
     internal class AdminConsoleViewModel : ViewModelBase
     {
         private User[] m_Users { get; set; }
+        private bool showDeletedUsers { get; set; }
         private KeyValuePair<string, string> m_ServerVersionInfo { get; set; }
 
         public User[]? Users
         {
             get { return m_Users; }
             set { m_Users = value; OnPropertyChanged(); }
+        }
+        public bool ShowDeletedUsers
+        {
+            get { return showDeletedUsers; }
+            set { showDeletedUsers = value; OnPropertyChanged(); }
         }
         public Array PermissionRoleEnumTypes
         {