+++ gamevault/ViewModels/GameDownloadViewModel.cs
@@ -1,4 +1,5 @@
-﻿using gamevault.Models;
+﻿using gamevault.Converter;
+using gamevault.Models;
 using gamevault.UserControls;
 using System;
 using System.Collections.Generic;
@@ -27,6 +28,7 @@ namespace gamevault.ViewModels
         private Visibility m_DownloadUIVisibility { get; set; }
         private Visibility m_ExtractionUIVisibility { get; set; }
         private Visibility m_DownloadFailedVisibility { get; set; }
+        private bool? createShortcut { get; set; }
 
         #endregion
 
@@ -101,11 +103,35 @@ namespace gamevault.ViewModels
             get { return m_IsDownloadResumed; }
             set { m_IsDownloadResumed = value; OnPropertyChanged(); }
         }
+        public bool? CreateShortcut
+        {
+            get
+            {
+                if (createShortcut == null)
+                {
+                    createShortcut = Preferences.Get(AppConfigKey.CreateDesktopShortcut, AppFilePath.UserFile) == "1";
+                }
+                return createShortcut;
+            }
+
+            set
+            {
+                createShortcut = value; OnPropertyChanged();
+                Preferences.Set(AppConfigKey.CreateDesktopShortcut, createShortcut == true ? "1" : "0", AppFilePath.UserFile);
+            }
+        }
         public string[] SupportedArchives
         {
             get
             {
-                return new string[] { ".7z", ".xz", ".bz2", ".gz", ".tar", ".zip", ".wim", ".ar", ".arj", ".cab", ".chm", ".cpio", ".cramfs", ".dmg", ".ext", ".fat", ".gpt", ".hfs", ".ihex", ".iso", ".lzh", ".lzma", ".mbr", ".msi", ".nsis", ".ntfs", ".qcow2", ".rar", ".rpm", ".squashfs", ".udf", ".uefi", ".vdi", ".vhd", ".vmdk", ".wim", ".xar", ".z" };
+                return Globals.SupportedArchives;
+            }
+        }
+        public Dictionary<GameType, string?> GameTypes
+        {
+            get
+            {
+                return Enum.GetValues(typeof(GameType)).Cast<GameType>().Where(v => v != GameType.UNDETECTABLE).ToDictionary(v => v, v => new EnumDescriptionConverter().Convert(v, null, null, null) as string);
             }
         }
     }